naver_land/requirements.txt
mysql-connector==2.2.9
Scrapy==2.5.1
